+++ srcs/Server_util.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Server_util.cpp                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: wluedara <wluedara@student.42.fr>          +#+  +:+       +#+        */
+/*   By: tkraikua <tkraikua@student.42.th>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/11 16:43:27 by csantivimol       #+#    #+#             */
-/*   Updated: 2024/01/20 02:55:37 by tkraikua         ###   ########.fr       */
+/*   Updated: 2024/01/21 19:32:26 by tkraikua         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,7 +64,7 @@ void Server::receive_message(int fd)
 	
 	recv(fd, buffer, sizeof(buffer), 0);
 	std::string text(buffer);
-	int found = text.find_first_of("\r\n");
+	unsigned long found = text.find_first_of("\r\n");
 	if (found != std::string::npos)
 		text = text.substr(0, found);
